+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - May 25th, 2021
+
+### Added
+
+- US Address generators
+- English name generators
+- YAML parser for easy serialization of large text arrays
+- Mutator for modifying existing generators
+
 ## [0.2.0] - May 24th, 2021
 
 ### Added